+++ drivers/gpu/drm/nouveau/nvkm/subdev/fb/gh100.c
@@ -14,7 +14,7 @@ gh100_fb_sysmem_flush_page_init(struct nvkm_fb *fb)
 	struct nvkm_device *device = fb->subdev.device;
 
 	// Ensure that the address is within hardware limits
-	WARN_ON(fb->sysmem.flush_page_addr > DMA_BIT_MASK(52));
+	WARN_ON((u64)fb->sysmem.flush_page_addr > DMA_BIT_MASK(52));
 
 	nvkm_wr32(device, NV_PFB_FBHUB_PCIE_FLUSH_SYSMEM_ADDR_HI, upper_32_bits(addr));
 	nvkm_wr32(device, NV_PFB_FBHUB_PCIE_FLUSH_SYSMEM_ADDR_LO, lower_32_bits(addr));
